simplify query logic
